api/requirements.txt
# django
django==3.1.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
factory-boy==3.0.1
Pillow==7.2.0
# django-debug-toolbar==2.2
# django-extensions==3.0.5
# uwsgi
uwsgi==2.0.18
# mysql
mysqlclient==1.4.6
# redis
redis==3.5.3
celery==4.4.7
django-filter==2.3.0
django-redis==4.12.1
